refactor(mathplug/): Consolidated module and module tests & modified setup file. Incermented version to v0.0.2.
