--- mathplug/add.py
@@ -1,6 +1,0 @@
-def add(*arg):
-    """Returns sum of numbers passed in."""
-    total = 0
-    for n in arg:
-        total += n
-    return total
