--- mathplug/square.py
@@ -1,3 +1,0 @@
-def square(arg):
-    """Returns the argument multiplied by itself."""
-    return arg * arg
