--- mathplug/square_root.py
@@ -1,5 +1,0 @@
-def square_root(arg):
-    """Returns the root of the argument."""
-    if arg < 0:
-        return None
-    return arg ** 0.5
